requirements.txt
Package           Version
----------------- -------
blinker           1.8.2
cachelib          0.13.0
click             8.1.7
colorama          0.4.6
Flask             3.0.3
Flask-Login       0.6.3
Flask-Session     0.8.0
Flask-SQLAlchemy  3.1.1
Flask-WTF         1.2.1
greenlet          3.0.3
itsdangerous      2.2.0
Jinja2            3.1.4
MarkupSafe        2.1.5
msgspec           0.18.6
pip               24.1.2
psycopg2-binary   2.9.9
python-dotenv     1.0.1
setuptools        70.1.1
SQLAlchemy        2.0.31
typing_extensions 4.12.2
Werkzeug          3.0.3
wheel             0.43.0
WTForms           3.1.2
